Handle failed docker builds

+++ src/main/php/xp/lambda/CreateRuntime.class.php
@@ -17,20 +17,22 @@ class CreateRuntime {
   public function run(): int {
     $docker= $this->command();
     $runtime= $this->image('runtime', $this->version, [], $this->rebuild);
+    if (null === $runtime) return 1;
 
+    // Verify runtime
     $code= "echo ' => PHP ', PHP_VERSION, ' & Zend ', zend_version(), ' @ ', php_uname(), PHP_EOL, ' => ', implode(' ', get_loaded_extensions()), PHP_EOL;";
     Console::writeLine();
     Console::writeLine("[+] Running {$runtime}\e[34m");
     passthru("{$docker} run --rm {$runtime} /opt/php/bin/php -r \"{$code}\"", $result);
     Console::writeLine("\e[0m");
 
+    // Extract runtime
     $container= uniqid();
     $commands= [
       "{$docker} create --name {$container} {$runtime}",
       "{$docker} cp {$container}:/opt/php/runtime.zip {$this->target}",
       "{$docker} rm -f {$container}",
     ];
-
     Console::writeLine('[+] Creating ', $this->target);
     foreach ($commands as $command) {
       Console::writeLinef("\e[34m => %s\e[0m", $command);
